refactor(objects): extract pack crate (#1100)
