+++ crates/objects/src/store/fs/mod.rs
@@ -11,7 +11,6 @@ mod pack_install_journal;
 #[cfg(test)]
 mod fs_tests;
 
-pub use fs_io::read_file_bytes_for_pack;
 pub use fs_store::{FsStore, LooseObjectWriteMode};
 pub use pack_install_journal::{
     DEFAULT_PACK_INSTALL_INTENT_TTL_SECS, PackInstallIntent, PackInstallMetricsSnapshot,
