+++ crates/pack/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "heddle-pack"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+description = "Heddle's pack and delta storage format."
+license.workspace = true
+repository.workspace = true
+keywords.workspace = true
+categories.workspace = true
+
+[dependencies]
+blake3.workspace = true
+bytes.workspace = true
+heddle-format = { path = "../format", version = "0.10" }
+heddle-fs-prims = { path = "../fs-prims", version = "0.10" }
+heddle-object-model = { path = "../object-model", version = "0.10" }
+memmap2.workspace = true
+tracing.workspace = true
+zstd = { workspace = true, optional = true }
+
+[features]
+zstd = ["dep:zstd", "heddle-format/zstd"]
+
+[dev-dependencies]
+tempfile.workspace = true
+
+[lib]
+path = "src/lib.rs"
+name = "heddle_pack"
