+++ crates/pack/src/lib.rs
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+//! Heddle's pack and delta storage format.
+
+pub use heddle_object_model::object;
+
+pub mod store;
