+++ crates/pack/src/store/mod.rs
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+//! Pack storage primitives.
+
+pub mod pack;
+
+pub use heddle_object_model::error::{HeddleError as StoreError, HeddleError, Result};
