+++ crates/pack/src/store/pack/shared.rs
