+++ crates/pack/src/store/pack/streaming_builder.rs
@@ -48,7 +48,7 @@
 //!   is a non-issue for the call site that motivated this builder.
 //! - **No path-grouped reordering.** Entries land in the order added.
 //! - **Output is a pack file at a path** rather than `(Vec<u8>, Vec<u8>)`.
-//!   Callers pair this with [`crate::store::ObjectStore::install_pack_from_path`]
+//!   Callers pair this with `objects::store::ObjectStore::install_pack_from_path`
 //!   which moves/installs the pack without copying it through RAM.
 //! - **Re-reads the pack at finalize** to compute the BLAKE3 trailer
 //!   checksum (the pack format hashes header+body, and the count goes
@@ -250,7 +250,8 @@ impl<W: Write + Read + Seek + SyncData> StreamingPackBuilder<W> {
         bucket_dir: PathBuf,
         declared_object_count: Option<u64>,
     ) -> Result<Self> {
-        crate::fs_atomic::create_dir_all_durable(&bucket_dir).map_err(StoreError::from)?;
+        heddle_fs_prims::fs_atomic::create_dir_all_durable(&bucket_dir)
+            .map_err(StoreError::from)?;
         let header_offset = pack_writer.stream_position().map_err(StoreError::from)?;
 
         // Write a placeholder header with `count = 0` unless the caller knows
@@ -653,7 +654,7 @@ impl<W: Write + Read + Seek + SyncData> StreamingPackBuilder<W> {
             .map_err(|e| StoreError::from(std::io::Error::other(e.to_string())))?;
         idx_file.sync_all().map_err(StoreError::from)?;
         if let Some(parent) = self.index_path.parent() {
-            crate::fs_atomic::sync_directory(parent).map_err(StoreError::from)?;
+            heddle_fs_prims::fs_atomic::sync_directory(parent).map_err(StoreError::from)?;
         }
         debug_assert_eq!(
             entries_written, self.object_count,
