+++ crates/pack/src/store/pack/varint.rs
