+++ scripts/check-facade-render-free.sh
@@ -9,6 +9,7 @@ manifests=(
   crates/objects/Cargo.toml
   crates/fs-prims/Cargo.toml
   crates/object-model/Cargo.toml
+  crates/pack/Cargo.toml
   crates/merge/Cargo.toml
   crates/semantic/Cargo.toml
   crates/refs/Cargo.toml
